+++ lib/aws_public_ips.rb
@@ -5,6 +5,7 @@ end
 
 require 'aws_public_ips/checks'
 require 'aws_public_ips/cli'
+require 'aws_public_ips/cli_options'
 require 'aws_public_ips/formatters'
 require 'aws_public_ips/utils'
 require 'aws_public_ips/version'