+++ spec/aws_public_ips/checks/redshift_spec.rb
@@ -5,8 +5,9 @@ describe ::AwsPublicIps::Checks::Redshift do
     stub_request(:post, 'https://redshift.us-east-1.amazonaws.com')
       .to_return(body: ::IO.read('spec/fixtures/redshift-classic-public.xml'))
 
-    expect(subject.run).to eq([
+    expect(run_check(subject)).to eq([
       {
+        region: 'us-east-1',
         id: 'classic',
         hostname: 'classic.csorkyt5dk7h.us-east-1.redshift.amazonaws.com',
         ip_addresses: %w[54.167.97.240 54.91.252.196 54.242.227.110]
@@ -21,8 +22,9 @@ describe ::AwsPublicIps::Checks::Redshift do
     stub_dns(
       'vpc-public-2.csorkyt5dk7h.us-east-1.redshift.amazonaws.com' => %w[54.156.7.130]
     )
-    expect(subject.run).to eq([
+    expect(run_check(subject)).to eq([
       {
+        region: 'us-east-1',
         id: 'vpc-public-2',
         hostname: 'vpc-public-2.csorkyt5dk7h.us-east-1.redshift.amazonaws.com',
         ip_addresses: %w[52.44.170.10 54.209.53.206 54.208.75.129 54.156.7.130]
@@ -34,12 +36,12 @@ describe ::AwsPublicIps::Checks::Redshift do
     stub_request(:post, 'https://redshift.us-east-1.amazonaws.com')
       .to_return(body: ::IO.read('spec/fixtures/redshift-vpc-private.xml'))
 
-    expect(subject.run).to eq([])
+    expect(run_check(subject)).to eq([])
   end
 
   it 'should handle clusters with a nil endpoint' do
     stub_request(:post, 'https://redshift.us-east-1.amazonaws.com')
       .to_return(body: ::IO.read('spec/fixtures/redshift-empty-endpoint.xml'))
-    expect { subject.run }.to raise_error(StandardError, /has a nil endpoint/)
+    expect { run_check(subject) }.to raise_error(StandardError, /has a nil endpoint/)
   end
 end