+++ spec/aws_public_ips/cli_spec.rb
@@ -1,27 +1,8 @@
 # frozen_string_literal: true
 
 describe ::AwsPublicIps::CLI do
-  it 'should parse the options' do
-    options = subject.parse(%w[--format json --services ec2,elb,redshift --verbose])
-    expect(options).to eq(format: 'json', services: %w[ec2 elb redshift], verbose: true)
-  end
-
-  it 'should raise on an invalid formatter' do
-    expect { subject.parse(%w[--format blah]) }.to raise_error(::ArgumentError, /Invalid format/)
-  end
-
-  it 'should raise on an invalid service' do
-    expect { subject.parse(%w[--service blah]) }.to raise_error(::ArgumentError, /Invalid service/)
-  end
-
-  it 'should select the right directory' do
-    ::Dir.chdir('/') do
-      options = subject.parse(%w[--service ec2 --format prettyjson])
-      expect(options).to include(services: %w[ec2], format: 'prettyjson')
-    end
-  end
-
   it 'should run' do
+    stub_describe_regions
     expect(::AwsPublicIps::Checks::Ec2).to receive(:run).and_return([{
       id: 'i-0f22d0af796b3cf3a',
       hostname: 'ec2-54-234-208-236.compute-1.amazonaws.com',
@@ -31,7 +12,17 @@ describe ::AwsPublicIps::CLI do
     subject.run(['-s', 'ec2'])
   end
 
+  it 'should run with progress' do
+    stub_describe_regions
+    stub_request(:post, 'https://ec2.us-east-1.amazonaws.com')
+      .to_return(body: ::IO.read('spec/fixtures/ec2.xml'))
+    expect(::STDOUT).to receive(:puts)
+    expect(::STDERR).to receive(:print).at_least(4).times
+    subject.run(['-p', '-s', 'ec2'])
+  end
+
   it 'should rescue exceptions' do
+    stub_describe_regions
     expect(subject).to receive(:check_service).and_raise(::StandardError)
     expect(::Process).to receive(:exit).with(1)
     expect(::STDERR).to receive(:puts)
@@ -39,12 +30,25 @@ describe ::AwsPublicIps::CLI do
   end
 
   it 'should print the version' do
+    stub_describe_regions
     expect(::STDOUT).to receive(:puts).with(::AwsPublicIps::VERSION)
-    subject.run(['--version'])
+    expect { subject.run(['--version']) }.to raise_error(::SystemExit)
   end
 
   it 'should print help' do
+    stub_describe_regions
     expect(::STDOUT).to receive(:puts)
-    subject.run(['--help'])
+    expect { subject.run(['--help']) }.to raise_error(::SystemExit)
+  end
+
+  it 'should print help on option exceptions' do
+    begin
+      stub_describe_regions
+      expectation = expect { subject.run(['--format', 'blah']) }
+      expectation.to output(/Invalid format/).to_stderr_from_any_process
+      expectation.to output(/Usage: /).to_stderr_from_any_process
+    rescue SystemExit => e
+      expect(e.status).to eq(1)
+    end
   end
 end